pyproject.toml
[tool.poetry]
name = "pandas-money"
version = "0.1.0"
description = "A Pandas/Pyarrow mashup for handling currency with Decimal128 performance"
authors = ["Rod Morison <rmorison@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "pandas_money"}]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.0.2"
pyarrow = "^12.0.0"
pandas-stubs = "^2.0.2.230605"
py-moneyed = "^3.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
mypy = "^1.3.0"
debugpy = "^1.6.7"
pylint = "^2.17.4"
python-lsp-server = "^1.7.3"
python-lsp-black = "^1.3.0"
pyls-isort = "^0.2.2"
isort = "^5.12.0"
pylsp-mypy = "^0.6.7"
flake8 = "^6.0.0"
pre-commit = "^3.3.2"
pytest = "^7.3.1"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["pandas_money"]

[tool.mypy]
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = false
check_untyped_defs = true
no_implicit_reexport = true

[tool.isort]
profile = "black"

[tool.black]
line-length = 88

[flake8]
# from https://black.readthedocs.io/en/stable/the_black_code_style/current_style.html#line-length
max-line-length = 88
select = ["C","E","F","W","B","B950"]
extend-ignore = ["E203", "E501", "W503"]

[tool.coverage.run]
omit =["*/tests/*", "tests.py"]
